Rename base.py to component.py
