+++ src/tribler/core/components/bandwidth_accounting/bandwidth_accounting_component.py
@@ -3,7 +3,7 @@ from tribler.core.components.bandwidth_accounting.community.bandwidth_accounting
     BandwidthAccountingTestnetCommunity,
 )
 from tribler.core.components.bandwidth_accounting.db.database import BandwidthDatabase
-from tribler.core.components.base import Component
+from tribler.core.components.component import Component
 from tribler.core.components.ipv8.ipv8_component import Ipv8Component
 from tribler.core.utilities.simpledefs import STATEDIR_DB_DIR
