+++ src/tribler/core/components/ipv8/ipv8_component.py
@@ -15,7 +15,7 @@ from ipv8.taskmanager import TaskManager
 
 from ipv8_service import IPv8
 
-from tribler.core.components.base import Component
+from tribler.core.components.component import Component
 from tribler.core.components.key.key_component import KeyComponent
 
 INFINITE = -1
