+++ src/tribler/core/components/key/tests/test_key_component.py
@@ -1,7 +1,5 @@
-import pytest
-
-from tribler.core.components.session import Session
 from tribler.core.components.key.key_component import KeyComponent
+from tribler.core.components.session import Session
 
 
 async def test_masterkey_component(tribler_config):
