+++ src/tribler/core/components/payout/payout_component.py
@@ -1,6 +1,6 @@
 from tribler.core import notifications
 from tribler.core.components.bandwidth_accounting.bandwidth_accounting_component import BandwidthAccountingComponent
-from tribler.core.components.base import Component
+from tribler.core.components.component import Component
 from tribler.core.components.ipv8.ipv8_component import Ipv8Component
 from tribler.core.components.payout.payout_manager import PayoutManager
 from tribler.core.components.reporter.reporter_component import ReporterComponent
