+++ src/tribler/core/components/reporter/exception_handler.py
@@ -7,7 +7,7 @@ from socket import gaierror
 from traceback import print_exception
 from typing import Callable, Optional
 
-from tribler.core.components.base import ComponentStartupException
+from tribler.core.components.component import ComponentStartupException
 from tribler.core.components.reporter.reported_error import ReportedError
 from tribler.core.sentry_reporter.sentry_reporter import SentryReporter
