+++ src/tribler/core/components/reporter/reporter_component.py
@@ -1,4 +1,4 @@
-from tribler.core.components.base import Component
+from tribler.core.components.component import Component
 from tribler.core.components.key.key_component import KeyComponent
 from tribler.core.components.reporter.exception_handler import default_core_exception_handler
 from tribler.core.utilities.unicode import hexlify
