+++ src/tribler/core/components/restapi/restapi_component.py
@@ -5,7 +5,7 @@ from ipv8.REST.root_endpoint import RootEndpoint as IPV8RootEndpoint
 
 from tribler.core.components.bandwidth_accounting.bandwidth_accounting_component import BandwidthAccountingComponent
 from tribler.core.components.bandwidth_accounting.restapi.bandwidth_endpoint import BandwidthEndpoint
-from tribler.core.components.base import Component, NoneComponent
+from tribler.core.components.component import Component, NoneComponent
 from tribler.core.components.gigachannel.gigachannel_component import GigaChannelComponent
 from tribler.core.components.gigachannel_manager.gigachannel_manager_component import GigachannelManagerComponent
 from tribler.core.components.ipv8.ipv8_component import Ipv8Component
