+++ src/tribler/core/components/restapi/tests/test_restapi_component.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock
 import pytest
 
 from tribler.core.components.bandwidth_accounting.bandwidth_accounting_component import BandwidthAccountingComponent
-from tribler.core.components.base import NoneComponent
+from tribler.core.components.component import NoneComponent
 from tribler.core.components.gigachannel.gigachannel_component import GigaChannelComponent
 from tribler.core.components.ipv8.ipv8_component import Ipv8Component
 from tribler.core.components.key.key_component import KeyComponent
