+++ src/tribler/core/components/tag/tag_component.py
@@ -1,5 +1,5 @@
 import tribler.core.components.metadata_store.metadata_store_component as metadata_store_component
-from tribler.core.components.base import Component
+from tribler.core.components.component import Component
 from tribler.core.components.ipv8.ipv8_component import Ipv8Component
 from tribler.core.components.key.key_component import KeyComponent
 from tribler.core.components.metadata_store.utils import generate_test_channels
