+++ src/tribler/core/utilities/tiny_tribler_service.py
@@ -4,7 +4,7 @@ import signal
 from pathlib import Path
 from typing import List, Optional
 
-from tribler.core.components.base import Component
+from tribler.core.components.component import Component
 from tribler.core.components.session import Session
 from tribler.core.utilities.osutils import get_root_state_directory
 from tribler.core.utilities.process_checker import ProcessChecker
